Extract code for calling hub.exe into HubService
