+++ GitPull.Services/IHubService.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Threading.Tasks;
+
+namespace GitPull.Services
+{
+    public interface IHubService
+    {
+        Task SyncRepositoryAsync(string solutionDir, Progress<string> progress);
+    }
+}
